feat(aeo): publish the ARD capability manifest - Task 5 of PLAN_branch_audit_and_pr
isitagentready.com scores this origin 93/100, Level 5. The only miss is
API/Auth/MCP & Skill Discovery at 7/8, for want of an ARD capability manifest.

The gap was real. This site already serves an MCP server card, an Agent Skills
index, an OpenAPI schema, a did:web identity, llms.txt and 492 Markdown twins --
and announced none of it in one place, so an agent had to already know each
convention to find each surface. The manifest is the index that makes the rest
discoverable.

Seven entries, every one a resource that returns 200 today. No A2A entry,
because there is no A2A agent: a capability manifest is a promise, and a
registry that indexes a 404 has been lied to. Half the new test suite exists to
enforce exactly that -- every advertised URL must resolve to a real source, the
headers must carry the CORS the spec requires, and robots.txt must not disallow
anything advertised.

That last check earned its keep immediately: /api/cfs-open.json sat under the
blanket Disallow: /api/ while being advertised as a capability. It has its own
Allow now. And the branch's own endpoint turned out to be missing from
openapi.json entirely, so it is documented there too.

Representative queries are written in both languages on purpose -- registries
embed them, and a manifest indexed only in English would be found only by
English speakers.

Verified live through wrangler over the production build: 200,
application/json, Access-Control-Allow-Origin: *, and all seven URLs 200.
925 -> 944 tests. The score itself cannot be confirmed until this deploys.
