+++ .wrangler/tmp/bundle-KekDG6/middleware-insertion-facade.js
@@ -1,8 +1,8 @@
-				import worker, * as OTHER_EXPORTS from "/app/.wrangler/tmp/pages-HuJOZK/functionsWorker-0.4757907487514158.mjs";
+				import worker, * as OTHER_EXPORTS from "/app/.wrangler/tmp/pages-2y4RVY/functionsWorker-0.5522104109476836.mjs";
 				import * as __MIDDLEWARE_0__ from "/home/node/.npm/_npx/d77349f55c2be1c0/node_modules/wrangler/templates/middleware/middleware-ensure-req-body-drained.ts";
 import * as __MIDDLEWARE_1__ from "/home/node/.npm/_npx/d77349f55c2be1c0/node_modules/wrangler/templates/middleware/middleware-miniflare3-json-error.ts";
 
-				export * from "/app/.wrangler/tmp/pages-HuJOZK/functionsWorker-0.4757907487514158.mjs";
+				export * from "/app/.wrangler/tmp/pages-2y4RVY/functionsWorker-0.5522104109476836.mjs";
 				const MIDDLEWARE_TEST_INJECT = "__INJECT_FOR_TESTING_WRANGLER_MIDDLEWARE__";
 				export const __INTERNAL_WRANGLER_MIDDLEWARE__ = [
