+++ public/_headers
@@ -3,11 +3,21 @@
 
 /
   Link: </.well-known/api-catalog>; rel="api-catalog"
+  Link: </.well-known/ai-catalog.json>; rel="ai-catalog"; type="application/json"
 
 /es/
   Link: </.well-known/api-catalog>; rel="api-catalog"
+  Link: </.well-known/ai-catalog.json>; rel="ai-catalog"; type="application/json"
 
 # Agent-readiness .well-known endpoints (RFC 9727, RFC 8414, RFC 9728, SEP-1649, agent-skills v0.2.0)
+
+# ARD capability manifest. `Access-Control-Allow-Origin: *` is required by the
+# spec, not optional: a registry indexes this from another origin.
+/.well-known/ai-catalog.json
+  Content-Type: application/json
+  Access-Control-Allow-Origin: *
+  Cache-Control: public, max-age=300, must-revalidate
+
 /.well-known/api-catalog
   Content-Type: application/linkset+json
   Cache-Control: public, max-age=300, must-revalidate
